+++ bin/lib/commands/build.js
@@ -1,0 +1,5 @@
+import * as util from '../util';
+
+export default (args) => {
+  util.buildFiles(util.CONSTANTS.rollupConfig);
+}
