+++ bin/lib/commands/index.js
@@ -1,0 +1,17 @@
+import newRepo from './new';
+import build from './build';
+import generate from './generate';
+import test from './test';
+import serve from './serve';
+import sketch from './sketch';
+import d from './delete';
+
+export {
+  newRepo,
+  build,
+  generate,
+  test,
+  serve,
+  sketch,
+  d
+};
